+++ example/components/custom-type-body.tsx
